Move thrift-derived parquet schema to public headers
