+++ cpp/src/io/parquet/parquet_gpu.hpp
@@ -18,13 +18,13 @@
 
 #include "error.hpp"
 #include "io/comp/comp.hpp"
-#include "io/parquet/parquet.hpp"
 #include "io/parquet/parquet_common.hpp"
 #include "io/statistics/statistics.cuh"
 #include "io/utilities/column_buffer.hpp"
 
 #include <cudf/detail/device_scalar.hpp>
 #include <cudf/io/datasource.hpp>
+#include <cudf/io/parquet_schema.hpp>
 #include <cudf/types.hpp>
 #include <cudf/utilities/span.hpp>
